Build app.
