+++ lib/Experiment.js
@@ -24,6 +24,10 @@ var _emitter = require("./emitter");
 
 var _emitter2 = _interopRequireDefault(_emitter);
 
+var _store = require("./store");
+
+var _store2 = _interopRequireDefault(_store);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -32,35 +36,11 @@ function _possibleConstructorReturn(self, call) { if (!self) { throw new Referen
 
 function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
-var store = void 0;
-
-var noopStore = {
-  getItem: function getItem() {},
-  setItem: function setItem() {}
-};
-
-if (typeof window !== 'undefined' && 'localStorage' in window && window['localStorage'] !== null) {
-  try {
-    var key = '__pushtell_react__';
-    window.localStorage.setItem(key, key);
-    if (window.localStorage.getItem(key) !== key) {
-      store = noopStore;
-    } else {
-      window.localStorage.removeItem(key);
-      store = window.localStorage;
-    }
-  } catch (e) {
-    store = noopStore;
-  }
-} else {
-  store = noopStore;
-}
-
 _emitter2.default.addActiveVariantListener(function (experimentName, variantName, skipSave) {
   if (skipSave) {
     return;
   }
-  store.setItem('PUSHTELL-' + experimentName, variantName);
+  _store2.default.setItem('PUSHTELL-' + experimentName, variantName);
 });
 
 var Experiment = function (_Component) {
@@ -89,7 +69,7 @@ var Experiment = function (_Component) {
       if (typeof activeValue === "string") {
         return activeValue;
       }
-      var storedValue = store.getItem('PUSHTELL-' + _this.props.name);
+      var storedValue = _store2.default.getItem('PUSHTELL-' + _this.props.name);
       if (typeof storedValue === "string") {
         _emitter2.default.setActiveVariant(_this.props.name, storedValue, true);
         return storedValue;
